requirements.txt
PyInstaller
PyQt5
pyserial
